feat: nhân bản sản phẩm, số mét lẻ, sửa menu & trang Danh mục
Tính năng:
- Nhân bản sản phẩm: copy cả ảnh, size và mọi thông số; bản sao để ẩn và mở
  thẳng trang sửa. Nút ở danh sách Sản phẩm và trong trang sửa.
- Vải mua được số lẻ: ô số lượng gõ tay được, nhận cả "2,2" lẫn "2.2".
  Nút − / + đưa về số nguyên gần nhất theo hướng bấm (1,3 → − về 1, + lên 2),
  ít nhất 1 mét. Giỏ hàng và thanh toán hiển thị "3,8 mét".
- Trang Danh mục có nút ↑ ↓ đổi thứ tự danh mục và danh mục con.
- Ô "Loại lụa" gộp với danh mục con thành một dropdown phân cấp: chọn xong tự
  gán trang vải tương ứng, khỏi điền 2 chỗ mà lệch nhau.
- Dòng "Chất liệu" cạnh ảnh vải giữ định dạng admin soạn (đậm/nghiêng, danh
  sách, ảnh) thay vì bị lột sạch HTML thành một dòng chữ dính liền.

Sửa lỗi:
- Sản phẩm vải không đặt hàng được: backend đem nhãn "2 mét" đi đối chiếu bảng
  size vốn không có bản ghi nào → luôn báo "Size ... hiện không có hàng".
- order_items.quantity là INTEGER nên không lưu nổi 3.8 → đổi sang NUMERIC(10,2).
- Trang Danh mục trong admin trắng trơn: thiếu khai báo const auth khiến
  template gọi auth.can() trên undefined (lỗi lọt từ đợt làm phân quyền).
- Mega menu "Sản phẩm" lúc mở lúc không: mouseenter không bắn lại sau khi bấm
  vào menu vì con trỏ chưa hề rời vùng hover → đổi sang mouseover.
- seed_categories.py ghi đè sort_order, chạy lại là xoá sạch thứ tự admin đã sắp.

Dữ liệu: migration 0013 đưa "Lụa tơ tằm" lên đầu menu một lần.

+++ backend/app/models/order.py
@@ -58,7 +58,8 @@ class OrderItem(BaseModel):
     product          = ForeignKeyField(Product, backref="order_items", null=True, on_delete="SET NULL")
     product_name     = CharField(max_length=255, null=True)   # snapshot
     size             = CharField(max_length=20)
-    quantity         = IntegerField(default=1)
+    # Vải bán theo mét nên cho số lẻ (VD 3.8 mét); quần áo vẫn là số nguyên
+    quantity         = DecimalField(max_digits=10, decimal_places=2, default=1)
     price            = DecimalField(max_digits=14, decimal_places=0)   # unit price snapshot (incl. adjustments)
     tailoring_method = CharField(max_length=60, null=True)   # option_key snapshot
     lining_type      = CharField(max_length=60, null=True)
